Extract slug resolution out of submissions.ts
Moves confirmSlugClaim/ensureSubmissionSlug/isSlugClaimed into
catalog/slug-resolver.ts as createSlugResolver, a small factory over
store and isSlugPublished. Closes out the Phase 3 Wave B seam
extraction series: the remaining bulk of submissions.ts is the
central dispatch/lifecycle/reconciler machinery, which doesn't
decompose cleanly via this pattern.
